lots more cleanup
