+++ crates/atuin-client/src/record/sync/builder.rs
@@ -5,7 +5,7 @@ use eyre::Result;
 use tracing::instrument;
 use typed_builder::TypedBuilder;
 
-use super::{SyncEngine, SyncError};
+use super::{DEFAULT_PAGE_SIZE, SyncEngine, SyncError};
 use crate::api_client::{Client, caps_client};
 use crate::record::sqlite_store::SqliteStore;
 use crate::settings::Settings;
@@ -60,6 +60,7 @@ impl SyncEngineInit<'_> {
         Ok(SyncEngine {
             client,
             store: self.store,
+            page_size: DEFAULT_PAGE_SIZE,
         })
     }
 }
