+++ crates/atuin/src/command/client/store/pull.rs
@@ -50,7 +50,8 @@ impl Pull {
             })
             .build()
             .connect()
-            .await?;
+            .await?
+            .with_page_size(self.page);
 
         let keyed = engine.keyed(&key);
         let (diff, _remote_index) = engine.diff().await?;
@@ -87,7 +88,7 @@ impl Pull {
             })
             .collect();
 
-        let (_, downloaded) = keyed.sync_remote(operations, self.page).await?;
+        let (_, downloaded) = keyed.sync_remote(operations).await?;
 
         println!("Downloaded {} records", downloaded.len());
